% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Title}
\usage{
check_duplicates(data_df, group_by_vars = NULL, print_dups = TRUE)
}
\arguments{
\item{data_df}{Dataframe to be checked for duplicates}

\item{group_by_vars}{A string or vector of strings of column name(s) to group by in asserting uniqueness}
}
\value{
Duplicate rows, both in general and by variable/column groupings
}
\description{
Title
}
\examples{
Coming Soon


}
