% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_values.R
\name{check_values}
\alias{check_values}
\title{A function to check to values from their from data sources}
\usage{
check_values(data_df, from, to = NULL, type, eval_func = as.numeric)
}
\arguments{
\item{data_df}{A dataframe that includes the from and to variables.}

\item{from}{A string or vector of strings that contain from variable names.}

\item{to}{A string or vector of strings containing the name of to variables (default is NULL).}

\item{type}{Either "map" or "cont" to specify the type of checking to be performed.}

\item{eval_func}{A function call for the type of check that should be performed on "cont" to variables (default is as.numeric).}
}
\value{
An invisible dataframe with the results of the check.
}
\description{
This function checks and inspects the mapping and values of specified columns
in a dataframe, comparing 'from' and 'to' columns if provided.
}
\examples{
# Assuming you have a DataFrame called 'my_data' with 'from_col' and 'to_col' columns
check_values(
  data_df = my_data,
  from = "from_col",
  to = "to_col",
  type = "map"
)
}
