% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dataframes.R
\name{compare_dataframes}
\alias{compare_dataframes}
\title{Compare 2 dataframes}
\usage{
compare_dataframes(
  oldData = NULL,
  newData = NULL,
  suffixTerm = "",
  indOutcomes = c(""),
  surgCol = "surgeon"
)
}
\arguments{
\item{oldData}{Old data}

\item{newData}{New data}

\item{suffixTerm}{Suffix term}

\item{indOutcomes}{indent outcomes (not sure if this is accurate)}

\item{surgCol}{surgeon column name}
}
\value{
final df list
}
\description{
Compare 2 dataframes
}
