% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_to_csv.R
\name{excel_to_csv}
\alias{excel_to_csv}
\title{Convert Excel File to CSV}
\usage{
excel_to_csv(path, filename, sheet = NULL)
}
\arguments{
\item{path}{Directory path containing Excel file}

\item{filename}{Excel filename with extension}

\item{sheet}{Sheet name or number. Default NULL reads first sheet}
}
\value{
Invisibly returns CSV output path
}
\description{
Convert Excel File to CSV
}
