% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islet-package.R
\docType{package}
\name{islet-package}
\alias{islet}
\alias{islet-package}
\title{islet: Provides Data Pipelining Functions}
\description{
This package houses data pipelining functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.shaunporwal.com/islet/}
}

}
\author{
\strong{Maintainer}: Shaun P. Porwal \email{shaun.Porwal@gmail.com}

}
\keyword{internal}
