% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islet.est.R
\name{islet.est.bp}
\alias{islet.est.bp}
\title{SigU_est' is already to be rendered.
Sig0_est' is already to be rendered.
llk' is already to be rendered.}
\usage{
islet.est.bp(ipc, datuse)
}
\description{
SigU_est' is already to be rendered.
Sig0_est' is already to be rendered.
llk' is already to be rendered.
}
