% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islet.test.R, R/islet.test_slope.R
\name{islet.lrt.unix}
\alias{islet.lrt.unix}
\title{SigU_est' is already to be rendered.
Sig0_est' is already to be rendered.
llk' is already to be rendered.}
\usage{
islet.lrt.unix(ipc, datuse, ktest)

islet.lrt.unix(ipc, datuse, ktest)
}
\description{
SigU_est' is already to be rendered.
Sig0_est' is already to be rendered.
llk' is already to be rendered.

SigU_est' is already to be rendered.
Sig0_est' is already to be rendered.
llk' is already to be rendered.
}
