% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISLET_main_fun.R
\name{ISLET}
\alias{ISLET}
\title{ISLET for regular tensor regression, 3-d case}
\usage{
ISLET(X, y, r1, r2, r3, p1, p2, p3)
}
\arguments{
\item{X}{input tensor, n*p1*p2*p3}

\item{y}{input response y}

\item{r}{input rank like r1,r2,r3}

\item{p}{input dimension like p1, p2, p3}
}
\value{
Estimate of coefficient tensor A.
}
\description{
This function do ISLET under the regular low-tucker rank tensor regression case. For detail about the procedure, see Zhang et al 2019.
}
\examples{
set.seed(2018)
n = c(1000); p = c(10); r = c(3); sig = c(5);
p1 = p2 = p3 = p
r1 = r2 = r3 = r
# Generate data
S.array = array(rnorm(r1*r2*r3), dim=c(r1,r2,r3)) # Core tensor
S = as.tensor(S.array)
E1 = matrix(rnorm(p1*r1), nrow = p1, ncol=r1)
E2 = matrix(rnorm(p2*r2), nrow = p2, ncol=r2)
E3 = matrix(rnorm(p3*r3), nrow = p3, ncol=r3)
X = as.tensor(array(rnorm(n*p1*p2*p3), dim=c(n,p1,p2,p3)))
# Parameter tensor
A = ttm(ttm(ttm(S, E1, 1), E2, 2), E3, 3)
eps = sig * rnorm(n) # Error
y = regression.product(X, A) + eps # response
snr <- sd(y)/sd(eps)
# Estimator
A_sketch <- sketch(X, y, r1, r2, r3, p1, p2, p3)
error <- fnorm(A_sketch$hatA - A)/ fnorm(A)
}
\references{
\insertRef{zhang2020islet}{ISLET}
}
