% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islet.est.R
\name{islet.solve.block}
\alias{islet.solve.block}
\title{SigU_est' is already to be rendered.
Sig0_est' is already to be rendered.
llk' is already to be rendered.}
\usage{
islet.solve.block(Y, datuse)
}
\description{
SigU_est' is already to be rendered.
Sig0_est' is already to be rendered.
llk' is already to be rendered.
}
