% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_log.R
\name{line_log}
\alias{line_log}
\title{Log Information About Comments and TODOs}
\usage{
line_log(
  type,
  item,
  current_file = get_current_file(),
  data_path = file.path(getwd(), "data", PROJECT_DICT$data_date)
)
}
\arguments{
\item{type}{A character string representing the type of annotation, either "COMMENT" or "TODO".}

\item{item}{A character string or object containing the comment or TODO line(s).}

\item{current_file}{A character string representing the name of the current file being edited. Default is the current file.}

\item{data_path}{A character string representing the path to the directory where log files should be saved. Default is a "secure_data" subfolder in the working directory.}
}
\value{
NULL. This function is called for its side effect of saving log files.
}
\description{
This function logs information about comments and TODOs added in a file, storing metadata
in a specified data path.
}
\examples{
# Log a comment
line_log("COMMENT", "This is a comment.")

# Log a TODO with a custom current_file and data_path
line_log("TODO", "This is a TODO.", "my_file.R", "custom_data_path")
}
