% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_banner.R
\name{make_banner}
\alias{make_banner}
\title{Create a Banner for Code Separation and Organization}
\usage{
make_banner(str_to_banner, banner_chr = "&", output = TRUE)
}
\arguments{
\item{str_to_banner}{A string that will be displayed in the center of the banner.}

\item{banner_chr}{A character to be used for creating the banner boundary. Default is '&'.}

\item{output}{A boolean indicating whether to print the banner to the console. Default is TRUE.}
}
\value{
An invisible string containing the generated banner.
}
\description{
This function generates a text banner for separating or organizing code.
The banner is also copied to the clipboard for easy pasting.
}
\examples{
\dontrun{
make_banner(str_to_banner = 'Section Title', banner_chr = '&', output = TRUE)
}
}
