% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_file.R
\name{run_file}
\alias{run_file}
\title{Function to run the "function"/"scripts" that populate main.R}
\usage{
run_file(
  file,
  args = NULL,
  ignore_warnings =
    "NAs introduced by coercion|failed to parse|no non-missing arguments to max; returning -Inf|no non-missing arguments to min; returning Inf"
)
}
\arguments{
\item{file}{A function/script call}

\item{args}{A list of parameters that should be specified prior to running the function/script specified in "file"}

\item{ignore_warnings}{For now, a default logic to muffle a few routine "warnings" that are likely best to prevent from clogging up the output in the main.R script (NAs values created).  Can set to "" to override this functionality.}
}
\value{
return_list
}
\description{
Wrapper function for a do.call() command applied to script function.  Benefit is de-clutter the calls.  Drawback is debugging more convoluted.
}
