% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_issue.R
\name{save_issue}
\alias{save_issue}
\title{Save issue data to a specified path}
\usage{
save_issue(
  data_df,
  rds,
  data_path = file.path(getwd(), "secure_data", PROJECT_DICT$data_date),
  issues_path = file.path(getwd(), "secure_data", PROJECT_DICT$data_date, "metadata",
    "issues")
)
}
\arguments{
\item{data_df}{A data frame containing issue data to be saved}

\item{rds}{An RDS object to be used for saving data}

\item{data_path}{An optional character string specifying the data path. Defaults to 'file.path(getwd(), "secure_data", PROJECT_DICT$data_date)'}

\item{issues_path}{An optional character string specifying the issues path. Defaults to 'file.path(getwd(), "secure_data", PROJECT_DICT$data_date, 'metadata', 'issues')'}
}
\value{
An invisible copy of the saved data frame
}
\description{
This function saves issue data to a specified path and prints the saved data to the console.
It takes a data frame, an RDS object, and optional data and issues paths as inputs.
}
\examples{
# Save issue data to the default paths
save_issue(data_df, rds)

# Save issue data to custom paths
save_issue(data_df, rds, "custom_data_path", "custom_issues_path")
}
