% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISLET_main_fun.R
\name{select_r}
\alias{select_r}
\title{Estimate Rank when rank is unknown- 3-d case}
\usage{
select_r(X, y, thres, r1, r2, r3, p1, p2, p3)
}
\arguments{
\item{X}{input tensor, n*p1*p2*p3}

\item{y}{input response y}

\item{thres}{thresholding in the autoselection of rank r, suggestion 1-2.}

\item{r}{input a conservative rank estimation r1_init,r2_init,r3_inits}

\item{p}{input dimension like p1, p2, p3}
}
\value{
Estimate Tucker rank of the parameter tensor
}
\description{
This function is used to estimate the rank r when the Tucker rank of tensor is unknown. For detail about the procedure, see reference Zhang et al 2019 and Zhang 2018 CROSS: Efficient low rank tensor completion.s
}
\examples{
# ISLET in the unknown rank setting
n = c(1000); p = c(10); r = c(3); sig = c(5);
p1 = p2 = p3 = p
r1 = r2 = r3 = r
thres = 1.5; # thresholding is greater than 1, and not too big. Suggestion is 1-2.
# data generate
A = ttm(ttm(ttm(S, E1, 1), E2, 2), E3, 3)
eps = sig * rnorm(n) # Error
y = regression.product(X, A) + eps # response
snr <- sd(y)/sd(eps)
# Get a estimate of rank
r1_init = r2_init = r3_init = 8
r_select <- select_r(X, y, thres, r1_init, r2_init, r3_init, p1, p2, p3)
# Do estimation
A_sketch <- sketch(X, y, r_select[1], r_select[2], r_select[3], p1, p2, p3)
#error
fnorm(A_sketch - A)/fnorm(A)
}
\references{
\insertRef{zhang2020cross}{ISLET}
}
