% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Check if there are duplicate rows in a dataframe based on grouped columns}
\usage{
check_duplicates(data_df, group_by_vars = NULL, print_dups = TRUE)
}
\arguments{
\item{data_df}{Dataframe to be checked for duplicates}

\item{group_by_vars}{A string or vector of strings of column name(s) to group by in asserting uniqueness}

\item{print_dups}{A boolean argument, if TRUE, will allow function to display duplicate rows}
}
\value{
Duplicate rows, both in general and by variable/column groupings
}
\description{
Check if there are duplicate rows in a dataframe based on grouped columns
}
\examples{
# Load the dplyr package
library(dplyr)

# Create a sample data frame
data_df <- data.frame(
  ID = c(1, 1, 2, 3, 4),
  Age = c(25, 25, 30, 40, 50),
  Income = c(50000, 50000, 60000, 70000, 80000)
)

# Call the function to check for duplicates in the entire data frame
check_duplicates(data_df)

# Call the function to check for duplicates within the "ID" column
check_duplicates(data_df, group_by_vars = "ID")

# Call the function to check for duplicates within the "ID" and "Age" columns
check_duplicates(data_df, group_by_vars = c("ID", "Age"))

}
