% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dataframes.R
\name{compare_dataframes}
\alias{compare_dataframes}
\title{Compare 2 dataframes}
\usage{
compare_dataframes(
  oldData = NULL,
  newData = NULL,
  suffixTerm = "",
  indOutcomes = c(""),
  surgCol = "surgeon"
)
}
\arguments{
\item{oldData}{A dataframe representing the old data.}

\item{newData}{A dataframe representing the new data. If NULL, the function will only analyze the oldData.}

\item{suffixTerm}{A character string representing the suffix term.}

\item{indOutcomes}{A character vector representing indent outcomes (not sure if this is accurate).}

\item{surgCol}{A character string representing the surgeon column name.}
}
\value{
A list of dataframes containing the results of the comparison.
}
\description{
This function compares two dataframes by analyzing their numeric, factor, character,
binary, and date fields. It returns a list of dataframes with the results of the comparison.
}
\examples{
old_data <- data.frame(...)
new_data <- data.frame(...)
compare_dataframes(oldData = old_data, newData = new_data)
}
