% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cases.R
\name{filter_cases}
\alias{filter_cases}
\title{Filter Cases from a Data Frame}
\usage{
filter_cases(
  data_df,
  select,
  logic,
  group_by_vars = NULL,
  rds,
  data_path = file.path(getwd(), "secure_data", PROJECT_DICT$data_date),
  remove = TRUE
)
}
\arguments{
\item{data_df}{A data frame to be filtered.}

\item{select}{A character vector of variable names used in the logical expression.}

\item{logic}{A character string representing the logical expression to filter cases.}

\item{group_by_vars}{A character vector of variable names to group by (default: NULL).}

\item{rds}{A character string specifying the filename for the RDS file storing removed cases.}

\item{data_path}{A character string specifying the file path for saving the RDS file (default: "secure_data/<data_date>").}

\item{remove}{A logical value indicating whether to remove cases that don't meet the logical expression (default: TRUE).}
}
\value{
An invisible data frame with either the cases that meet the logical expression (if remove = TRUE) or the ungrouped input data (if remove = FALSE).
}
\description{
This function filters cases from a data frame based on the given logical expression.
It can also group the data by certain variables if needed.
}
\examples{
# Filter cases where age is greater than or equal to 18
filtered_data <- filter_cases(data_df = my_data,
select = "age",
logic = "age >= 18",
rds = "cases_removed")
}
