% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_win_to_mac.R
\name{from_win_to_mac}
\alias{from_win_to_mac}
\title{Title}
\usage{
from_win_to_mac(windows_path, mounted_drive_name)
}
\arguments{
\item{windows_path}{Windows path to be converted to Mac path}

\item{mounted_drive_name}{Mounted drive to be included in path}
}
\value{
Converted Mac path
}
\description{
Title
}
\examples{
Coming soon!
}
