% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_structure.R
\name{llm_structure}
\alias{llm_structure}
\title{Structure Data using LLM}
\usage{
llm_structure(df, col_name, prompt)
}
\arguments{
\item{df}{A dataframe.}

\item{col_name}{Name of the column in df to be processed.}

\item{prompt}{Prompt to be combined with each cell's value.}
}
\value{
Dataframe with a new column suffixed with "_llm".
}
\description{
Applies an LLM to structure data from a specified column.
}
