% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_banner.R
\name{make_banner}
\alias{make_banner}
\title{Create a Banner for Code Separation/Organization}
\usage{
make_banner(str_to_banner, banner_chr = "&", output = TRUE)
}
\arguments{
\item{str_to_banner}{String to Convert to Banner}

\item{banner_chr}{Banner Boundary Character}

\item{output}{Boolean, TRUE to output, FALSE to Not Output}
}
\value{
Invisible String
}
\description{
Create a Banner for Code Separation/Organization
}
\examples{
make_banner(str_to_banner = 'asdf', banner_chr = '&', output = TRUE)
}
