% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_function.R
\name{parse_function}
\alias{parse_function}
\title{Parse a dataframe and compute statistics for different variable types}
\usage{
parse_function(
  parse_df,
  suffixTerm = "",
  indOutcomes = c(""),
  surgCol = "surgeon",
  addYears = FALSE
)
}
\arguments{
\item{parse_df}{Dataframe to parse}

\item{suffixTerm}{Suffix term (default: "")}

\item{indOutcomes}{A character vector of individual outcomes to compare per surgeon (default: c(""))}

\item{surgCol}{A character string specifying the surgeon column (default: "surgeon")}

\item{addYears}{A boolean indicating whether to add year values (default: FALSE)}
}
\value{
A list of dataframes containing the computed statistics for the different variable types
}
\description{
This function takes a dataframe as input and computes statistics for
different variable types, including dates, binary, characters, factors,
and numeric variables. It can also compute statistics per surgeon and
per year if specified.
}
\examples{
# Assuming you have a dataframe called df
parsed_data <- parse_function(df)
# To include year values and individual outcomes per surgeon:
parsed_data <- parse_function(df, indOutcomes = c("outcome1", "outcome2"), addYears = TRUE)
}
