% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_function.R
\name{parse_function}
\alias{parse_function}
\title{Parsing function}
\usage{
parse_function(
  parse_df,
  suffixTerm = "",
  indOutcomes = c(""),
  surgCol = "surgeon",
  addYears = FALSE
)
}
\arguments{
\item{parse_df}{Dataframe to parse}

\item{suffixTerm}{Suffix term}

\item{indOutcomes}{Indent outcomes?}

\item{surgCol}{Surgeon Column}

\item{addYears}{Add years}
}
\value{
Result Dataframe
}
\description{
Parsing function
}
