% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_data.R
\name{read_raw_data}
\alias{read_raw_data}
\title{Read and Clean Raw Data}
\usage{
read_raw_data(file, col_caps = TRUE, str_caps = TRUE)
}
\arguments{
\item{file}{A string specifying the path to the CSV file to be read.}

\item{col_caps}{Logical. Whether to convert column names to uppercase. Default is TRUE.}

\item{str_caps}{Logical. Whether to convert string column values to uppercase. Default is TRUE.}
}
\value{
A data frame with cleaned column names and data.
}
\description{
This function reads a CSV file and performs several cleaning steps:
\enumerate{
\item Optionally converts column names to uppercase using janitor::clean_names.
\item Attempts to parse character columns as dates in ymd format.
\item Optionally converts character columns to uppercase if not parsed as dates.
}
}
