% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sql_query.R
\name{to_sql_query}
\alias{to_sql_query}
\title{Format text from SQL file into proper SQL query}
\usage{
to_sql_query(filepath)
}
\arguments{
\item{filepath}{Path to SQL file}
}
\value{
SQL query string
}
\description{
Format text from SQL file into proper SQL query
}
\examples{
Coming Soon!
}
